requirements.txt
asgiref==3.8.1
astroid==3.3.6
colorama==0.4.6
dill==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==3.3.2
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
tomlkit==0.13.2
tzdata==2024.2
uritemplate==4.1.1
weasyprint==61.2
openpyxl==3.1.2
kafka-python==2.0.2
confluent-kafka==2.8.2
django-oauth-toolkit==2.3.0
oauthlib==3.2.2
django-allauth==0.61.1
python-social-auth==0.3.6
social-auth-app-django==5.4.0
requests-oauthlib==1.3.1
